requirements.txt
datasets
fasttext-wheel
keras
nltk
numpy
scikit-learn
tensorflow